STYLE: Force consistent formatting
The vtkReader.cxx vtkReader.h files are in 3 different locations
in the ukftractography dirctory.  All locations need to be fixed
but the formatting of each file is inconsistent.  By
applying a fixed consistent formatting (ITK style) it is easier
to ensure that the files are truely consistent and redundancies can
be removed.

This step will ease the comparision to code that can be moved into
the common static library directory.