STYLE: Refactor common code in UKFCommonLib
Redundant naming of files and classes within same
named files did not have the same code.  This caused
many difficulties in understanding what the code was
doing, and it required redundant fixes to the code
when a problem was found.

After refactoring, all tests continue to pass so
it is assumed that the refactoring steps so far
are working properly.