BUG: VTK file format was invalid
According to http://www.vtk.org/VTK/img/file-formats.pdf

- "#vtk DataFile Version 3.0"
+"# vtk DataFile Version 3.0"
       ^
       \ NOTICE THE EXTRA SPACE HERE!
         That space is REQUIRED for valid format!

The first part is the file version and identifier. This part contains
the single line: "# vtk DataFile Version x.x".  This line must be **exactly**
as shown with the exception of the version number x.x, which will vary
with different releases of VTK. (Note: the current version number is
3.0. Version 1.0 and 2.0 files are compatible with version 3.0 files.)