ENH: Respect thread limits from environment
Setting the ITK number of threads must be done
carefully when used as part of the Slicer environment.

When this is used as a shared library module inside of slicer
it would reset **ALL** of Slicers ITK functions and programs
to be set to the number of threads specified by UKF after
UKF was run.   The solution is to reset the ITK number of threads
when exiting the program.  This is accomplished with a object that
does the resetting upon deletion.

As a side benefit, when UKF is used inside of a batch processing
system (i.e. SGE, OGS, PBS, torque, etc...) it will respect the
number of threads allocated by the scheduler.