PERF: Simplify code to ease maintenance, and performance
The code was unnecessarily complex with member variables
that were only used in one location and were hard coded
to constant values in all cases.  By making them constant
values and co-locating their definition to the member function
that used them it greatly simplifies understanding the
intent of the code.