BUG: replace baseline files with correct ones.
I enhanced the CompareFibers program to a) use VTK readers
and b) traverse the vtkPolyData objects returned and
actually compare all the fiber tracking data.

This actually broke the tests, because the files
generated by the current version of UKFTractography
has different data in them than the existing
baselines.  The actual fiber points are the same,
but the 'decorative' data is different.