COMP: Check for UKFTractography_BUILD_SLICER_EXTENSION
According to JC, this is set if the Slicer/Extensions/CMake
build method is used to build an extension, so check for
that to suppress things only needed in standalone superbuild.

COMP: dont override install/output paths when building Slicer extension

Actually suppress doing this unless it's a standalone
SuperBuild.

COMP: git rid of abolute paths for install targests

I'm not sure why these are in the build files copied
from NamicExternalProjects -- I imagine they should
not be there in any environment.