COMP: Consistently find teem library
Since the teem library provides a TeemConfig.cmake file if it is either
build as part of Superbuild or install on the system, it is sufficient
to use the "find_package()" call.

This commit fixes a configure error [1] by removing the extra call to
"find_library" that ended up finding the teem library installed on the
system.

More specifically, in the Slicer case, the Teem_DIR was used by the
previous call to "find_package(teem)" to find the teem library built by
Slicer.

Modules ended up referencing a mix of system teem library and Slicer teem
library.

[1] Configure rrror fixed by this commit:

8<----8<----8<----8<----8<----8<----8<----8<----
Cannot generate a safe linker search path for target ConvertVTK because
files in some directories may conflict with libraries in implicit
directories:

  link library [libteem.so] in /usr/lib may be hidden by files in:
    /home/jchris/Projects/Slicer-2-SuperBuild-Release/teem-build/bin
8<----8<----8<----8<----8<----8<----8<----8<----