COMP: Simplify management of output and install directories
This commit remove the Slicer specific settings from SlicerExectionModel
and ensure the project will be configured with the expected values. Then,
it is not needed anymore to overwrite the SlicerExecutionModel_DEFAULT_*
variable in the current scope.

This has been tested for both a SlicerExtension and a standalone built.
In both case the libraries are built where expected.