BUG: fixing some bugs in computedispersion.cxx
Fixing some bugs in computedispersion.cxx that probably resulted from typos.
In addition, I changed the line if(inDiskPlaneCount > 0) to > 10
for speed up. There is no need to perform the computations inside this
if statement if inDiskPlaneCount < 10, as the function will return false
in this case - see the following if statement.