COMP: Improved propagation of Superbuild VARS
Fix UKF build for Slicer:
UKF was not being placed in the proper directory
for Slicer, so attempts were made to fix the build
process for Slicer.

install TARGETS given no RUNTIME DESTINATION for executable target

In order to build stand-alone, as a slicer extension, and as
part of other build systems, the variables for
SEM CLI building must be provided through the SuperBuild variables.

Updated to VTK 6 from slicer Prefer to
use git hash from  Slicer to improve
consistency with Slicer module.

Syncronize with Slicer: Try to set the
external package dependancies to be very
similar to Slicer to maximize overlap of
consistent build proceedures.  This is
done in the hopes that future maintenance
can be minimized.

ITK was updated to avoid HDF5 conflicts between VTK
and ITK H5detect programs.