PERF: Use Eigen's built in DiagonalMatrix
This allows Eigen to use optimized computations
when it can determine that the matrix is a diagonal
matrix.

A further benefit of this optimization is that it
avoids creation of new memory allocation per call to
diffusion_euler because the correctly formated values
are passed in directly.

100% tests passed, 0 tests failed out of 29
Total Test time (real) =   0.80 sec