Diffusion Propagator model added
In this commit, the Diffusion Propagator model is added. It is a
multi-tensor bi-exponential model, with free water estimation. The
following changes were made:

- 5 flags have been added to the command line parser
(--diffusionPropagator, --minRTOP, --maxNMSE, --maxUKFIterations and
--recordRTOP) (UKFTractography.xml, UKFTractography.cxx)

- The model is implemented in filter_Simple2BiExp_FW.cc /
filter_Simple2BiExp_FW.h

- filter_Simple2BiExp_FW.cc has been added to the CMakeLists.txt

- The VTK writer has been updated to take into account the recorded
measures, which are the return-to-origin probabilites (RTOP), instead of
FA and TRACE.

- UnpackTensor has been modified so that, when estimating the single
tensor in the case of the diffusionPropagator model, only the gradients
with a b-value between 800 and 1500 s/mm² are taken into account.
Moreover, the neperien logarithm is not applied directly on the signal
anymore, since we need the signal latter on in the initialization, and
it is more computationally expensive to interpolate again than storing
the log_e(signal) in a small vector.

- When estimating the initial seed state, the LBFGSB optimizer from ITK
is used for the non-linear least square fitting. It is also possible to
iterate the UKF at the starting point to estimate the initial seed
state, simply by uncommenting the call to the function InitLoopUKF, and
by setting the desired number of iterations. However we found with
Yogesh that the nl-LSQ fitting was giving better results.

- The Step2T function has been modified to allow multiple iterations of
the UKF, when using the diffusionPropagator model.

- Whith this model, when following a fiber, negative free water will
simply causes the fiber to stop instead of making the program exit.