COMP: ITKv5 compatibility for new threading models
ITKv5 re-organized the threading capabilities to allow
new threading models.

The itkPlatformMultiThreader is the new class that provides
a multithreader class that provides support for multithreaded
execution using Windows or POSIX threads.
This class can be used to execute a single
method on multiple threads, or to specify a method per thread.