STYLE: Simplify VTK external project removing unneeded custom build step
This custom build command was originally introduced in Slicer r15695 to
support building VTK wrapping against a custom installation of python.
Since the UKF package does not rely on python wrapping, it can be removed.

For reference, the custom step was also removed in the Slicer version
of the VTK external project in Slicer r23072.

References:
* http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=15695
* http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=23072